Add files via upload
# TCS iON RiO 45 Internship

Sentiment analysis refers to identifying as well as classifying the sentiments that are expressed in the text source. Tweets are often useful in generating a vast amount of sentiment data upon analysis. These data are useful in understanding the opinion of the people about a variety of topics
## Problem Statement
In this project, I try to implement a Twitter sentiment analysis model that helps to overcome the challenges of identifying the sentiments of the tweets. 
The necessary details regarding the dataset are:
    The Sentiment140 Dataset which consists of 1,600,000 tweets that have been extracted using the Twitter API
    target: the polarity of the tweet (positive or negative)
    ids: Unique id of the tweet
    date: the date of the tweet
    flag: It refers to the query. If no such query exists then it is NO QUERY.
    user: It refers to the name of the user that tweeted
    text: It refers to the text of the tweet
## Machine learning Pipeline
The various steps involved in the Machine Learning Pipeline are :

Import Necessary Dependencies
Read and Load the Dataset
Exploratory Data Analysis
Data Visualization of Target Variables
Data Preprocessing
Splitting our data into Train and Test Subset
Transforming Dataset using TF-IDF Vectorizer
Function for Model Evaluation
Model Building
Conclusion